[WIP][EasyAsync] Make it easier to extend batch class
