+++ packages/EasyAsync/src/Events/Batch/BatchCancelledEvent.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EonX\EasyAsync\Events\Batch;
+
+final class BatchCancelledEvent extends AbstractBatchEvent
+{
+    // No body needed.
+}
